order_system/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.19
django-rest-framework==0.1.0
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==23.0.0
inotify==0.2.10
kombu==5.4.2
nose==1.3.7
packaging==24.2
prompt_toolkit==3.0.50
psycopg==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
